// src/app/shared/models/Course.ts
export class Course {
  id: string;
  name: string;
  course_code: string;
  price_tag: string;
  image_path: string;
  total_time: number;
  total_lessons: number;
  first_lesson: string;
  description: string;
  level: string;
  viewers: number;
  created_at: string;
  update_at: string;

  constructor() {
  }
}